+++ raytracer/src/aarect.rs
@@ -8,16 +8,16 @@ use crate::vec3::Vec3;
 
 #[derive(Clone)]
 pub struct XYRect<M: Material> {
-    pub x0: f64,
-    pub x1: f64,
-    pub y0: f64,
-    pub y1: f64,
-    pub k: f64,
+    pub x0: f32,
+    pub x1: f32,
+    pub y0: f32,
+    pub y1: f32,
+    pub k: f32,
     pub material: M,
 }
 
 impl<M: Material> XYRect<M> {
-    pub fn new(x0: f64, x1: f64, y0: f64, y1: f64, k: f64, material: M) -> Self {
+    pub fn new(x0: f32, x1: f32, y0: f32, y1: f32, k: f32, material: M) -> Self {
         Self {
             x0,
             x1,
@@ -30,7 +30,7 @@ impl<M: Material> XYRect<M> {
 }
 
 impl<M: Material> Hittable for XYRect<M> {
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<AxisAlignedBoundingBox> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<AxisAlignedBoundingBox> {
         let outputbox = AxisAlignedBoundingBox::new(
             Vec3::new(self.x0, self.y0, self.k - 0.0001),
             Vec3::new(self.x1, self.y1, self.k + 0.0001),
@@ -38,7 +38,7 @@ impl<M: Material> Hittable for XYRect<M> {
         Some(outputbox)
     }
 
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let t = (self.k - ray.origin().z()) / ray.direction().z();
         if t < t_min || t > t_max {
             return None;
@@ -78,16 +78,16 @@ impl<M: Material> Hittable for XYRect<M> {
 
 #[derive(Clone)]
 pub struct XZRect<M: Material> {
-    pub x0: f64,
-    pub x1: f64,
-    pub z0: f64,
-    pub z1: f64,
-    pub k: f64,
+    pub x0: f32,
+    pub x1: f32,
+    pub z0: f32,
+    pub z1: f32,
+    pub k: f32,
     pub material: M,
 }
 
 impl<M: Material> XZRect<M> {
-    pub fn new(x0: f64, x1: f64, z0: f64, z1: f64, k: f64, material: M) -> Self {
+    pub fn new(x0: f32, x1: f32, z0: f32, z1: f32, k: f32, material: M) -> Self {
         Self {
             x0,
             x1,
@@ -100,7 +100,7 @@ impl<M: Material> XZRect<M> {
 }
 
 impl<M: Material> Hittable for XZRect<M> {
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<AxisAlignedBoundingBox> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<AxisAlignedBoundingBox> {
         let outputbox = AxisAlignedBoundingBox::new(
             Vec3::new(self.x0, self.k - 0.0001, self.z0),
             Vec3::new(self.x1, self.k + 0.0001, self.z1),
@@ -108,7 +108,7 @@ impl<M: Material> Hittable for XZRect<M> {
         Some(outputbox)
     }
 
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let t = (self.k - ray.origin().y()) / ray.direction().y();
         if t < t_min || t > t_max {
             return None;
@@ -145,8 +145,8 @@ impl<M: Material> Hittable for XZRect<M> {
         })
     }
 
-    fn pdf_value(&self, origin: Vec3, direction: Vec3, wavelength: f64) -> f64 {
-        if let Some(rec) = self.hit(&Ray::new(origin, direction, 0.0, wavelength), 0.001, f64::INFINITY) {
+    fn pdf_value(&self, origin: Vec3, direction: Vec3, wavelength: f32) -> f32 {
+        if let Some(rec) = self.hit(&Ray::new(origin, direction, 0.0, wavelength), 0.001, f32::INFINITY) {
             let area = (self.x1 - self.x0) * (self.z1 - self.z0);
             let distance_squared = rec.t * rec.t * direction.length_squared();
             let cosine = direction.dot(&rec.normal).abs() / direction.length();
@@ -159,9 +159,9 @@ impl<M: Material> Hittable for XZRect<M> {
 
     fn random(&self, origin: Vec3) -> Vec3 {
         let random_point = Vec3::new(
-            rand::thread_rng().gen_range::<f64>(self.x0, self.x1),
+            rand::thread_rng().gen_range::<f32>(self.x0, self.x1),
             self.k,
-            rand::thread_rng().gen_range::<f64>(self.z0, self.z1),
+            rand::thread_rng().gen_range::<f32>(self.z0, self.z1),
         );
         random_point - origin
     }
@@ -169,16 +169,16 @@ impl<M: Material> Hittable for XZRect<M> {
 
 #[derive(Clone)]
 pub struct YZRect<M: Material> {
-    pub y0: f64,
-    pub y1: f64,
-    pub z0: f64,
-    pub z1: f64,
-    pub k: f64,
+    pub y0: f32,
+    pub y1: f32,
+    pub z0: f32,
+    pub z1: f32,
+    pub k: f32,
     pub material: M,
 }
 
 impl<M: Material> YZRect<M> {
-    pub fn new(y0: f64, y1: f64, z0: f64, z1: f64, k: f64, material: M) -> Self {
+    pub fn new(y0: f32, y1: f32, z0: f32, z1: f32, k: f32, material: M) -> Self {
         Self {
             y0,
             y1,
@@ -191,7 +191,7 @@ impl<M: Material> YZRect<M> {
 }
 
 impl<M: Material> Hittable for YZRect<M> {
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<AxisAlignedBoundingBox> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<AxisAlignedBoundingBox> {
         let outputbox = AxisAlignedBoundingBox::new(
             Vec3::new(self.k - 0.0001, self.y0, self.z0),
             Vec3::new(self.k + 0.0001, self.y1, self.z1),
@@ -199,7 +199,7 @@ impl<M: Material> Hittable for YZRect<M> {
         Some(outputbox)
     }
 
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let t = (self.k - ray.origin().x()) / ray.direction().x();
         if t < t_min || t > t_max {
             return None;
